requirements.txt
flask
tensorflow
numpy
gunicorn
pymysql==1.1.0
python-dotenv